+++ src/Validation/carName.js
@@ -1,6 +1,4 @@
 const MAX_NAME_LENGTH = 5;
-const MIN_GAME_COUNT = 0;
-const MAX_GAME_COUNT = 100;
 
 export const validateCarsNameLength = (input) => {
   if (input.split(",").some((carName) => carName.length > MAX_NAME_LENGTH)) {
@@ -23,20 +21,5 @@ export const validateDuplicatedCarName = (input) => {
   const origin = input.split(",");
   if (target.length !== origin.length) {
     throw new Error("[Error] 중복된 자동차 이름은 안됩니다.");
-  }
-};
-
-export const validateGameCountType = (input) => {
-  // TODO: '' 잡아내기
-  const isInteger = Number.isInteger(Number(input));
-  if (isInteger === false) {
-    throw new Error("[Error] 시도 횟수는 소수,문자열,NaN,Infinity 안됩니다.");
-  }
-};
-
-export const validateGameCountRange = (input) => {
-  const gameCount = Number(input);
-  if (gameCount <= MIN_GAME_COUNT || gameCount >= MAX_GAME_COUNT) {
-    throw new Error("[Error] 시도 횟수는 0이하, 100이상 안됩니다.");
   }
 };
